+++ lib/src/parser/ledger_lines/ledger_line.dart
@@ -1,5 +1,7 @@
 import '../../core/core.dart';
 
+typedef LedgerLineStreamProvider = Future<Stream<LedgerLine>> Function(String path);
+
 // Abstract parent for all LedgerLines, i.e. a parsed line in a ledger file
 abstract class LedgerLine {
 }
@@ -67,13 +69,22 @@ class NoteLine extends LedgerLine {
 // Represents an account line in a ledger file, i.e. "account ...."
 class AccountLine extends LedgerLine {
   final String name;
-  AccountLine(this.name);
+  final List<String> matchers;
+  AccountLine(this.name, {this.matchers = const []});
 
   @override
-  String toString() => "AccountLine(name: $name)";
+  String toString() => "AccountLine(name: $name, matchers: $matchers)";
+
+  bool matchersEquals(List<String> otherMatchers) {
+    if (otherMatchers.length != matchers.length) return false;
+    for (int i = 0; i < matchers.length; i += 1) {
+      if (matchers[i] != otherMatchers[i]) return false;
+    }
+    return true;
+  }
 
   @override
-  bool operator ==(Object other) => (other is AccountLine) && (name == other.name);
+  bool operator ==(Object other) => (other is AccountLine) && (name == other.name) && matchersEquals(other.matchers);
 
   @override
   int get hashCode => name.hashCode;