+++ test/parser/line_to_edit_transformer_test.dart
@@ -17,12 +17,19 @@ const testDataAdjustingBlankLine = """
     Expenses:Music       \$ 20
 """;
 
+void onTransformError(exc, stackTrace) => print("Transform error: $exc\n$stackTrace");
+
 void main() {
   group('line to edit transformer', () {
     test('transform strings in stream', () async {
       final stringStream = Stream<String>.fromIterable(testData1.split("\n"));
       final ledgerLineStream = stringStream.transform(LedgerStringToLineTransformer(onTransformError: (obj, stackTrace) => print(obj)));
-      final ledgerEditsStream = ledgerLineStream.transform(LedgerLineToEditsTransformer());
+      final ledgerEditsStream = ledgerLineStream.transform(
+          LedgerLineToEditsTransformer(
+            streamForIncludedFileCallback: (path) => Future.value(Stream.empty()),
+            onTransformError: onTransformError
+          ),
+      );
       final ledgerEdits = await ledgerEditsStream.toList();
       expect(ledgerEdits, [
         LedgerEditAddAccount('Assets:Checking'),
@@ -38,7 +45,7 @@ void main() {
     test('handle posting with no amount', () async {
       final stringStream = Stream<String>.fromIterable(testDataAdjustingBlankLine.split("\n"));
       final ledgerLineStream = stringStream.transform(LedgerStringToLineTransformer(onTransformError: (obj, stackTrace) => print(obj)));
-      final ledgerEditsStream = ledgerLineStream.transform(LedgerLineToEditsTransformer());
+      final ledgerEditsStream = ledgerLineStream.transform(LedgerLineToEditsTransformer(streamForIncludedFileCallback: (path) => Future.value(Stream.empty()), onTransformError: onTransformError));
       final ledgerEdits = await ledgerEditsStream.toList();
       expect(ledgerEdits, [
         LedgerEditAddAccount('Assets:Checking'),